tests/requirements.txt
# AlphaWatch Dependencies
anthropic>=0.34.0
pyyaml>=6.0
requests>=2.31.0
yfinance>=0.2.40
schedule>=1.2.0          # Local scheduler (dev)
python-dotenv>=1.0.0     # .env support
pytest>=8.0.0
